# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bank of Blockchain API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: Get all users (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersListResponse"
  /users/{id}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Get user by ID (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetail"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /users/{id}/activate:
    post:
      operationId: activateUser
      tags: [users]
      summary: Activate user account (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /users/{id}/suspend:
    post:
      operationId: suspendUser
      tags: [users]
      summary: Suspend user account (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [users]
      summary: Get admin dashboard stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        country:
          type: string
        dateOfBirth:
          type: string
        role:
          type: string
          enum: [user, admin]
        status:
          type: string
          enum: [pending, active, suspended]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - firstName
        - lastName
        - role
        - status
        - createdAt
    UserDetail:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            balances:
              $ref: "#/components/schemas/Balances"
            bankAccounts:
              type: array
              items:
                $ref: "#/components/schemas/BankAccount"
    UpdateUserRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        country:
          type: string
        status:
          type: string
          enum: [pending, active, suspended]
        role:
          type: string
          enum: [user, admin]
    UsersListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - users
        - total
        - page
        - limit
    Balances:
      type: object
      properties:
        eur:
          type: number
        usd:
          type: number
        btc:
          type: number
      required:
        - eur
        - usd
        - btc
...
    BankAccount:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        beneficiaryName:
          type: string
        iban:
          type: string
        bic:
          type: string
        bankName:
          type: string
        country:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - beneficiaryName
        - iban
        - createdAt
...
    AdminStats:
      type: object
      properties:
        totalUsers:
          type: integer
        activeUsers:
          type: integer
        pendingUsers:
          type: integer
        suspendedUsers:
          type: integer
        totalTransactions:
          type: integer
        pendingTransactions:
          type: integer
        completedTransactions:
          type: integer
        rejectedTransactions:
          type: integer
        totalBankTransfers:
          type: integer
        totalCryptoTransfers:
          type: integer
        pendingBankTransfers:
          type: integer
        pendingCryptoTransfers:
          type: integer
      required:
        - totalUsers
        - activeUsers
        - pendingUsers
        - suspendedUsers
        - totalTransactions
        - pendingTransactions
        - completedTransactions
        - rejectedTransactions
        - totalBankTransfers
        - totalCryptoTransfers
        - pendingBankTransfers
        - pendingCryptoTransfers